Reorganise changes/removals to be per-entity
We extract the oldest acked tick from all buffers, and use this
to collect a list of changed entity candidates, where a candidate
contains the entity, and the list of components changed and
removed, with the respective ticks.

This is then filtered per-client to discard any candidates with
change ticks older than the specific client's last acked tick.

Packet format unchanged for now.

This is preliminary work - this will lead to manual diff
fragmentation (and rooms/priority work), by allowing us to include
all change and removal data for an entity in one place in the
packet, and stop when the packet is full.